AudioSTT/requirements.txt
openai-whisper==20231117
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.4

